requirements/requirements.txt
alembic==1.13.1
alembic-postgresql-enum==1.3.0
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
black==24.4.2
boto3==1.35.51
boto3-stubs==1.35.51
botocore==1.35.51
botocore-stubs==1.35.51
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==42.0.7
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.115.4
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
jmespath==1.0.1
Mako==1.3.3
MarkupSafe==2.1.5
mypy-boto3-cloudformation==1.35.41
mypy-boto3-dynamodb==1.35.24
mypy-boto3-ec2==1.35.48
mypy-boto3-lambda==1.35.49
mypy-boto3-rds==1.35.50
mypy-boto3-s3==1.35.46
mypy-boto3-sqs==1.35.0
mypy-extensions==1.0.0
openai==1.33.0
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.1.18
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.3.3
pydantic_core==2.18.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.41.2
tqdm==4.66.4
types-awscrt==0.23.0
types-s3transfer==0.10.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
